% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{aggr_and_rep}
\alias{aggr_and_rep}
\title{Aggregate and repeat}
\usage{
aggr_and_rep(x, fun = mean)
}
\arguments{
\item{x}{(ts) Any time series}

\item{fun}{(function) Aggregation function, mean by default}
}
\value{
description
}
\description{
Helper function to repeat the aggregate annual value of a series on
each period.
}
\details{
Applies \code{aggregate.ts} to the series to get the annual values
and then repeats those values for every subyear period.
}
\examples{
aggr_and_rep(gdp_volume) |> plot()
}
