% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{aggr_rep_lag}
\alias{aggr_rep_lag}
\title{Aggregate, repeat and lag}
\usage{
aggr_rep_lag(x, fun = mean, k = 0)
}
\arguments{
\item{x}{(ts) Any time series}

\item{fun}{(function) Aggregation function, mean by default}

\item{k}{(int) Units to lag.}
}
\value{
description
}
\description{
Helper function to repeat the aggregate annual value of a series on
each period, and possibly lag it.
}
\details{
Applies \code{aggregate.ts} to the series to get the annual values
and then repeats those values for every subyear period.

The \code{k} parameter is passed to \code{stats::lag}.
}
\examples{
aggr_rep_lag(gdp_volume) |> plot()
}
