% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{apply_multivariate}
\alias{apply_multivariate}
\title{Apply method to multivariate}
\usage{
apply_multivariate(x, f, ...)
}
\arguments{
\item{x}{A multivariate time series of class "mts".}

\item{f}{A function that takes an univariate series as input.}

\item{...}{Arguments for \code{f}.}
}
\description{
This function applies a function for univariate series ("ts") to a multivariate
("mts").
}
