% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{chain_from_pyp_q}
\alias{chain_from_pyp_q}
\title{Get chain-linked indices}
\usage{
chain_from_pyp_q(x, ref_year, normalize = T)
}
\arguments{
\item{x}{A pyp series of class "ts".}

\item{ref_year}{Reference year ("num") for the chain-linked series.}

\item{normalize}{Make reference year = 100.}
}
\value{
The chain-linked series of class "ts".
}
\description{
Computes the quarterly chain-linked index series from a pyp series.
}
