% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{chain_vol_from_iq}
\alias{chain_vol_from_iq}
\title{Get chain-linked volume measure}
\usage{
chain_vol_from_iq(iq_chain, current, ref_year)
}
\arguments{
\item{iq_chain}{Chain-linked volume indices.}

\item{current}{Current prices.}

\item{ref_year}{Reference year for the chain-linked indices.}
}
\value{
The chain-linked volume measures.
}
\description{
Returns chain-linked volume given chain-linked volume indices and current prices.
}
