% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{change_ref_year_uni}
\alias{change_ref_year_uni}
\title{Change reference year (univariate)}
\usage{
change_ref_year_uni(x, new_ref)
}
\arguments{
\item{x}{An univariate chain-linked series (with annual overlap) of class "ts".}

\item{new_ref}{New reference year. Must be such that \code{start(x) <=
new_ref <= end(x)}.}
}
\value{
The re-referenced index series of class "ts".
}
\description{
Changes the reference year of an univariate chain-linked series (with annual
overlap).
}
