% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{check_reference}
\alias{check_reference}
\title{Checks the reference period}
\usage{
check_reference(x, ref_period, tol = 0.01)
}
\arguments{
\item{x}{A chain-linked series of indices of class "ts".}

\item{ref_period}{Reference period to be tested.}

\item{tol}{Tolerance}
}
\value{
TRUE if the reference period is c
}
\description{
Takes the mean of the values in the reference period and checks if
it is equal to 100.
}
