% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_gr}
\alias{compute_gr}
\title{Compute the growth-rate series}
\usage{
compute_gr(x, s)
}
\arguments{
\item{x}{(ts) A time series.}

\item{s}{(int) Lag at which the growth-rate is computed.}
}
\value{
Series of growth-rates.
}
\description{
Function that computes the growth-rate series of a given time series.
}
\examples{
compute_gr(gdp_current, 4)
}
