% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{concat_references}
\alias{concat_references}
\title{Concatenate two reference years}
\usage{
concat_references(x1, x2, tol = 0.01)
}
\arguments{
\item{x1}{First section of the series.}

\item{x2}{Second section of the series.}

\item{tol}{Tolerance for checking the overlapping period.}
}
\description{
Takes two overlapping sections of an index series, each with a
different reference year, and outputs a joined series with reference
that of the first section.
}
\examples{
x <- ts(sample(98:103, 20, TRUE), start = 2001, frequency = 4) |>
  chain_from_pyp(2001)
x1 <- window(x, end = 2004) |> change_ref_year(2002)
x2 <- window(x, start = 2003) |> change_ref_year(2004)
y <- concat_references(x1, x2) |> change_ref_year(2001)
plot(x)
lines(x1, col = "green")
lines(x2, col = "red")
points(y, col = "blue")

}
