% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp_current}
\alias{gdp_current}
\title{Spanish GDP (Current prices)}
\format{
A univariate time series object.
}
\source{
{https://ine.es/jaxiT3/Tabla.htm?t=67823&L=1} Spanish National Statistics Institute.
}
\description{
Spanish GDP from 1995 Q1 to 2024 Q4.
}
