% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp_volume}
\alias{gdp_volume}
\title{Spanish GDP (Chain-linked volume)}
\format{
A univariate time series object
}
\source{
{https://ine.es/jaxiT3/Tabla.htm?t=67824&L=1} Spanish National Statistics Institute.
}
\description{
Quantity chain-linked indices of the Spanish GDP from 1995 Q1 to 2024 Q4 with reference year 2020.
}
