% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{get_pyp}
\alias{get_pyp}
\title{Get pyp indices}
\usage{
get_pyp(x, x_a = NULL)
}
\arguments{
\item{x}{(ts) Chain-linked series with annual overlap.}

\item{x_a}{(ts) Annual chain-linked series.}
}
\value{
The pyp series of class "ts".
}
\description{
Computes the pyp index series from a chain-linked series.
}
\examples{
get_pyp(gdp_es_volume)

}
