% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{get_q_index}
\alias{get_q_index}
\title{Get quantity index}
\usage{
get_q_index(current, constant)
}
\arguments{
\item{current}{Current prices series of class "ts".}

\item{constant}{Constant prices (pyp) series of class "ts".}
}
\value{
List of time series ("ts"): the quarterly IPs and IQs (pyp).
}
\description{
Returns the series of quantity indices in previous year prices from
a current prices and
}
\examples{
gdp_es_pyp <- get_pyp(gdp_es_volume)
gdp_es_constant <- gdp_es_current / gdp_es_constant * 100
get_q_index(gdp_es_current, gdp_es_constant)

}
