% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_v_index}
\alias{get_v_index}
\title{Get value index}
\usage{
get_v_index(current)
}
\arguments{
\item{current}{(ts) Series of current prices series.}
}
\value{
(ts) Series of value indices.
}
\description{
Returns the (not chain-linked) series of value indices from a series
of current prices.
}
\details{
The value of the resulting series x at (y,s), where y is the year and s
is the subyear period, is current(y,s)/current(y)
}
\examples{
get_v_index(gdp_current)

}
