% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexNumberTools-package.R
\docType{package}
\name{IndexNumberTools-package}
\alias{IndexNumberTools}
\alias{IndexNumberTools-package}
\title{IndexNumberTools: Working with Index Numbers}
\description{
A set of utilities for manipulating index numbers series including chain-linking, re-referencing, and computing growth rates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mserrano-ine.github.io/IndexNumberTools/}
  \item \url{https://github.com/mserrano-ine/IndexNumberTools}
}

}
\author{
\strong{Maintainer}: Miguel Serrano \email{miguel.serrano.martin@ine.es}

}
\keyword{internal}
