% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{iq_pyp_from_money}
\alias{iq_pyp_from_money}
\title{Get volume index from current and pyp prices}
\usage{
iq_pyp_from_money(current, constant)
}
\arguments{
\item{current}{Current prices series of class "ts".}

\item{constant}{Constant prices (pyp) series of class "ts".}
}
\value{
List of time series ("ts"): the quarterly IPs and IQs (pyp).
}
\description{
Returns the series of pyp volume indices given current prices
and pyp prices.
}
