% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{iv_pyp_from_money}
\alias{iv_pyp_from_money}
\title{Get value index from current prices}
\usage{
iv_pyp_from_money(current)
}
\arguments{
\item{current}{Current prices series of class "ts".}
}
\value{
the series of IVs (pyp).
}
\description{
Returns the series of pyp value indices given current prices.
}
