% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{pyp_from_chain}
\alias{pyp_from_chain}
\title{Get pyp indices}
\usage{
pyp_from_chain(x)
}
\arguments{
\item{x}{A chain-linked series (with annual overlap) of class "ts".}
}
\value{
The pyp series of class "ts".
}
\description{
Computes the pyp index series from a chain-linked series.
}
