% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{pyp_from_chain_uni}
\alias{pyp_from_chain_uni}
\title{Get pyp indices (univariate)}
\usage{
pyp_from_chain_uni(x)
}
\arguments{
\item{x}{A chain-linked series (with annual overlap) of class "ts".}
}
\value{
The pyp series of class "ts".
}
\description{
Computes the pyp index series from an univariate chain-linked series.
}
