% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{pyp_from_money}
\alias{pyp_from_money}
\title{Get index from current and constant prices}
\usage{
pyp_from_money(current, constant)
}
\arguments{
\item{current}{Current prices series of class "ts".}

\item{constant}{Constant prices (pyp) series of class "ts".}
}
\value{
List of time series ("ts"): the quarterly IPs and IQs (pyp).
}
\description{
Get index from current and constant prices
}
