% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apply_to_columns}
\alias{apply_to_columns}
\title{Apply method to multivariate}
\usage{
apply_to_columns(x, f, ...)
}
\arguments{
\item{x}{(mts) A multivariate time series.}

\item{f}{(function) A function that takes an univariate series as input.}

\item{...}{Arguments for \code{f}.}
}
\description{
This function applies a function for univariate series ("ts") to a
multivariate series ("mts").
}
