% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{change_ref_year}
\alias{change_ref_year}
\title{Change reference year}
\usage{
change_ref_year(x, new_ref)
}
\arguments{
\item{x}{A chain-linked series (with annual overlap) of class "ts".}

\item{new_ref}{New reference year. Must be such that \code{start(x) <=
new_ref <= end(x)}.}
}
\value{
The re-referenced index series of class "ts".
}
\description{
Changes the reference year of a chain-linked series (with annual
overlap).
}
\examples{
change_ref_year(gdp_es_volume, 2015)
plot(gdp_es_volume)
lines(change_ref_year(gdp_es_volume, 2015))

}
