% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{change_ref_year}
\alias{change_ref_year}
\title{Change reference year}
\usage{
change_ref_year(x, new_ref_year)
}
\arguments{
\item{x}{(ts) A chain-linked series with annual overlap.}

\item{new_ref_year}{(num) New reference year.}
}
\value{
The re-referenced index series.
}
\description{
Changes the reference year of a chain-linked series (with annual
overlap).
}
\examples{
change_ref_year(gdp_volume, 2015)
plot(gdp_volume)
lines(change_ref_year(gdp_volume, 2015))

}
