% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{change_ref_year}
\alias{change_ref_year}
\title{Change reference year}
\usage{
change_ref_year(x, new_ref)
}
\arguments{
\item{x}{A chain-linked series (with annual overlap) of class "ts".}

\item{new_ref}{New reference year. Must be such that \code{start(x) <=
new_ref <= end(x)}.}
}
\value{
The re-referenced index series of class "ts".
}
\description{
Changes the reference year of a chain-linked series (with annual
overlap).
}
\examples{
set.seed(23)
x <- sample(95:105, 12, TRUE) |>
  ts(start = 2005, frequency = 4) |>
  chain_from_pyp(2006)
aggregate(x, FUN = mean)
y <- change_ref_year(x, 2007)
aggregate(y, FUN = mean)
set.seed(23)
x <- sample(95:105, 24, TRUE) |>
  matrix(ncol = 2) |>
  ts(start = 2005, frequency = 4) |>
  chain_from_pyp(2006)
aggregate(x, FUN = mean)
y <- change_ref_year(x, 2007)
aggregate(y, FUN = mean)
}
