% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{get_q_index}
\alias{get_q_index}
\title{Get quantity index}
\usage{
get_q_index(current, constant)
}
\arguments{
\item{current}{(ts) Values at current prices.}

\item{constant}{(ts) Values at previous year prices.}
}
\value{
Series of quantity indices for previous year prices.
}
\description{
Returns the series of quantity indices in previous year prices from
a current prices and
}
\examples{
gdp_pyp <- get_pyp(gdp_volume)
gdp_constant <- gdp_current / gdp_pyp * 100
get_q_index(gdp_current, gdp_constant)
}
