% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexNumberTools-package.R
\docType{package}
\name{IndexNumberTools-package}
\alias{IndexNumberTools}
\alias{IndexNumberTools-package}
\title{IndexNumberTools: Working With Index Numbers}
\description{
Chain-linking, base shifting, computing pyp indices, etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mserrano-ine.github.io/IndexNumberTools/}
  \item \url{https://github.com/mserrano-ine/IndexNumberTools}
}

}
\author{
\strong{Maintainer}: Miguel Serrano \email{miguel.serrano.martin@ine.es}

}
\keyword{internal}
