% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexNumberTools-package.R
\docType{package}
\name{IndexNumberTools-package}
\alias{IndexNumberTools}
\alias{IndexNumberTools-package}
\title{IndexNumberTools: Tools For Working With Index Numbers}
\description{
Tools for transforming chain-linked to pyp and vice versa, changing the reference year, etc.
}
\keyword{internal}
